pyproject.toml
[project]
name = "birdclef-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "torch>=2.3.0",
    "torchaudio>=2.3.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "timm>=1.0.0",
    "audiomentations>=0.36.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "wandb>=0.17.0",
    "kaggle>=1.6.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]
